chore: rename NewCollectionPages to CategoryPages for clarity
